close #1 close #2 Added Gaussian and Mol to main adt code
