--- opt/DFT_MOL/equigeom.dat
@@ -1,3 +1,0 @@
-0.0000000000	0.0000000000	-0.3241438387
-0.0000000000	1.1040126191	0.1418637008
-0.0000000000	-1.1040126191	0.1418637008
