--- opt/DFT_MOL/frequency.dat
@@ -1,1 +1,0 @@
-757.3300000000	1408.5200000000	1732.9000000000
