--- opt/DFT_MOL/geom.dat
@@ -1,3 +1,0 @@
-N  0.000000000  0.000000000  -0.324503807
-O  0.000000000  1.104583330  0.142043685
-O  0.000000000  -1.104583330  0.142043685
