--- opt/DFT_MOL/geom.xyz
@@ -1,5 +1,0 @@
-3
-geometry file created for optimization
-N,0.000000000,0.000000000,-0.324503807
-O,0.000000000,1.104583330,0.142043685
-O,0.000000000,-1.104583330,0.142043685
