--- opt/DFT_MOL/optg_mol.inp
@@ -1,14 +1,0 @@
-
-memory,1GB
-
-basis=cc-pVDZ
-
-symmetry,nosym
-
-geometry=geom.xyz
-
-{ks,b3lyp}
-optg;
-{frequencies;print,hessian}
-
----
