--- opt/DFT_MOL/wilson.dat
@@ -1,9 +1,0 @@
-0.0000000000	0.0000000000	-0.0000000000
-0.0000000000	0.0000000000	0.2170300000
-0.1656300000	0.1490800000	-0.0000000000
--0.0000000000	-0.0000000000	0.0000000000
-0.1182700000	-0.1313900000	-0.0950000000
--0.0725000000	-0.0652600000	-0.0401000000
--0.0000000000	-0.0000000000	0.0000000000
--0.1182700000	0.1313900000	-0.0950000000
--0.0725000000	-0.0652600000	0.0401000000
