+++ test_runs/optimization/gaussian.config
@@ -5,8 +5,8 @@ memory = 1000,m
 processor = 2
 method = ks,b3lyp
 basis  = cc-pVDZ
-; spin   = 2
-; charge = 0
+spin   = 2
+charge = 0
 symmetry = nosymm
