+++ test_runs/optimization/mol.config
@@ -1,7 +1,7 @@
-#Configuration file for the optimization and frequency by Gaussian package of NO2
+#Configuration file for the optimization and frequency by Molpro package of NO2
 
 [optInfo]
-memory = 1GB
+memory = 500,m
 processor = 1
 method = ks,b3lyp
 basis  = cc-pVDZ
